add res.status and create middleware with Joi
