+++ models/contacts.json
@@ -1,1 +1,1 @@
-[{"id":"1","name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":"2","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"3","name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":"4","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"5","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"6","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"7","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"8","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"9","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"10","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"name":"Ivan","email":"111@ukr.net","phone":"3434-3432-34324","id":"f134eddf-b9e5-47b1-bed9-7b7c9143d814"},{"name":"Ivan","email":"111@ukr.net","phone":"3434-3432-34324","id":"d64f7b7b-5b24-4231-ba2c-fe82bc7576f3"},{"name":"Ivan","email":"111@ukr.net","phone":"3434-3432-34324","id":"dd432ea4-6cd8-4af9-8a30-5c825f0a933e"},{"name":"Barbarossa","email":"barba@gmail.com","phone":"(067)-4543-4353-435","id":"1f4ecec4-f39f-47cc-8a5b-c40ed876e282"},{"name":"Timur","email":"timur@ukr.net","phone":"45345354","id":"bf1d8ba7-1e81-4dd5-aa54-db41bab48a77"},{"name":"fdrsh","email":"timur@ukr.net","phone":"45345354sdsd","id":"0516c8e4-a09a-4138-bb58-5148484d91f8"},{"name":"Bondar","email":"Bondar@roz.com","phone":"5555555","id":"2565f8de-42d2-4c95-9c9e-23e08c98c4ff"},{"name":"fdrsh","email":"timur@ukr.net","phone":"45345354","id":"8b97d587-a368-477b-8c77-b1612ade3ce0"}]
+[{"id":"1","name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":"2","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"3","name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":"4","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"5","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"6","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"7","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"8","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"9","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"10","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"name":"Ivan","email":"111@ukr.net","phone":"3434-3432-34324","id":"f134eddf-b9e5-47b1-bed9-7b7c9143d814"},{"name":"Ivan","email":"111@ukr.net","phone":"3434-3432-34324","id":"d64f7b7b-5b24-4231-ba2c-fe82bc7576f3"},{"name":"Ivan","email":"111@ukr.net","phone":"3434-3432-34324","id":"dd432ea4-6cd8-4af9-8a30-5c825f0a933e"},{"name":"Barbarossa","email":"barba@gmail.com","phone":"(067)-4543-4353-435","id":"1f4ecec4-f39f-47cc-8a5b-c40ed876e282"},{"name":"Timur","email":"timur@ukr.net","phone":"45345354","id":"bf1d8ba7-1e81-4dd5-aa54-db41bab48a77"},{"name":"fdrsh","email":"timur@ukr.net","phone":"45345354sdsd","id":"0516c8e4-a09a-4138-bb58-5148484d91f8"},{"name":"Bolivar","email":"Bondar@roz.com","phone":"34324234","id":"2565f8de-42d2-4c95-9c9e-23e08c98c4ff"},{"name":"fdrsh","email":"timur@ukr.net","phone":"45345354","id":"8b97d587-a368-477b-8c77-b1612ade3ce0"}]
