+++ routes/api/contacts.js
@@ -1,24 +1,12 @@
 const express = require("express");
-const Joi = require("joi");
+
 const contactsOperations = require("../../models/contacts");
 const router = express.Router();
-
-const schema = Joi.object({
-  name: Joi.string().alphanum().min(3).max(30).required(),
-
-  phone: Joi.string()
-    // eslint-disable-next-line prefer-regex-literals
-    .pattern(new RegExp("^[0-9]{3,30}$")),
-
-  email: Joi.string().email({
-    minDomainSegments: 2,
-    tlds: { allow: ["com", "net"] },
-  }),
-});
+const { schema } = require("../../middleware/scema");
 
 router.get("/", async (req, res, next) => {
   const allContacts = await contactsOperations.listContacts();
-  return res.json({
+  return res.status(200).json({
     status: 200,
     data: {
       result: allContacts,
@@ -29,14 +17,14 @@ router.get("/", async (req, res, next) => {
 router.get("/:contactId", async (req, res, next) => {
   const contact = await contactsOperations.getContactById(req.params.contactId);
   if (contact) {
-    return res.json({
+    return res.status(200).json({
       status: 200,
       data: {
         result: contact,
       },
     });
   } else {
-    return res.json({
+    return res.status(404).json({
       status: 404,
       message: "Not found",
     });
@@ -46,19 +34,21 @@ router.get("/:contactId", async (req, res, next) => {
 router.post("/", async (req, res, next) => {
   const validationResult = schema.validate(req.query);
   if (validationResult.error) {
-    return res.json({ status: 400, message: validationResult.error });
+    return res
+      .status(400)
+      .json({ status: 400, message: validationResult.error });
   }
   const { name, email, phone } = req.query;
   if (name && email && phone) {
     const contact = await contactsOperations.addContact({ name, email, phone });
-    return res.json({
+    return res.status(201).json({
       status: 201,
       data: {
         result: contact,
       },
     });
   } else {
-    return res.json({
+    return res.status(400).json({
       status: 400,
       message: "missing required name field",
     });
@@ -68,12 +58,12 @@ router.post("/", async (req, res, next) => {
 router.delete("/:contactId", async (req, res, next) => {
   const contact = await contactsOperations.removeContact(req.params.contactId);
   if (contact) {
-    return res.json({
+    return res.status(200).json({
       status: 200,
       message: "contact deleted",
     });
   } else {
-    return res.json({
+    return res.status(404).json({
       status: 404,
       message: "Not found",
     });
@@ -83,7 +73,7 @@ router.delete("/:contactId", async (req, res, next) => {
 router.put("/:contactId", async (req, res, next) => {
   const { name, email, phone } = req.query;
   if (!name || !email || !phone) {
-    return res.json({
+    return res.status(400).json({
       status: 400,
       message: "missing fields",
     });
@@ -91,7 +81,9 @@ router.put("/:contactId", async (req, res, next) => {
 
   const validationResult = schema.validate(req.query);
   if (validationResult.error) {
-    return res.json({ status: 400, message: validationResult.error });
+    return res
+      .status(400)
+      .json({ status: 400, message: validationResult.error });
   }
 
   const addedContact = await contactsOperations.updateContact(
@@ -99,12 +91,12 @@ router.put("/:contactId", async (req, res, next) => {
     { name, email, phone }
   );
   if (addedContact) {
-    return res.json({
+    return res.status(200).json({
       status: 200,
       result: addedContact,
     });
   } else {
-    return res.json({
+    return res.status(404).json({
       status: 404,
       message: "Not found",
     });
